Rename chart.json to charts.json to match other apps
