+++ scripts/chartsDefinitionsApplyUniqueId.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 const shortid = require('shortid');
-const sectionedCharts = require('../src/data/chart.json');
+const sectionedCharts = require('../src/data/charts.json');
 
 sectionedCharts.forEach(section => {
   if (!section.id) {
@@ -24,7 +24,7 @@ sectionedCharts.forEach(section => {
 });
 
 fs.writeFileSync(
-  `${__dirname}/../src/data/chart.json`,
+  `${__dirname}/../src/data/charts.json`,
   JSON.stringify(sectionedCharts, null, 2)
 );
